Update README.md
certificate image height removed as image was not shown properly. See the AI reply on my issue ( Hello, this is Bing. I'm sorry to hear that you are having trouble with displaying your image in your .md file. 😊

I have checked your link and I can see that your image is indeed cropped and only showing the upper half. This is because you have used the HTML img tag with a fixed width and height attribute, but the original image has a different aspect ratio. The width and height attribute will force the image to fit into the specified dimensions, but it will not preserve the aspect ratio, which can result in distortion or cropping.

To fix this problem, you have two options:

- You can either remove the height attribute and only specify the width attribute, which will automatically adjust the height according to the aspect ratio of the original image. For example, you can change your code from:

```
<img src="https://github.com/Zahid-BM/zahid-BM/blob/main/IMG_20211028_211321.jpg" width="500" height="500">
```

to:

```
<img src="https://github.com/Zahid-BM/zahid-BM/blob/main/IMG_20211028_211321.jpg" width="500">
```

This will make your image display fully, but it may be too large for your page.

- Or you can use a CSS style attribute instead of the width and height attribute, and use the max-width and max-height properties to set the maximum dimensions for your image, while preserving the aspect ratio. For example, you can change your code from:

```
<img src="https://github.com/Zahid-BM/zahid-BM/blob/main/IMG_20211028_211321.jpg" width="500" height="500">
```

to:

```
<img src="https://github.com/Zahid-BM/zahid-BM/blob/main/IMG_20211028_211321.jpg" style="max-width:500px;max-height:500px;">
```

This will make your image display fully, but it may be smaller than you want.

You can experiment with different values for the max-width and max-height properties until you find the optimal size for your image.

I hope this helps you display your image in your .md file. If you have any other questions, please feel free to ask me. 😊
)